+++ Package.swift
@@ -3,8 +3,8 @@ import PackageDescription
 let package = Package(
     name: "SwiftyBot",
     dependencies: [
-        .Package(url: "https://github.com/vapor/vapor.git", majorVersion: 1, minor: 3),
-        .Package(url: "https://github.com/FabrizioBrancati/BFKit-Swift.git", majorVersion: 2, minor: 1)
+        .Package(url: "https://github.com/vapor/vapor.git", majorVersion: 1, minor: 5),
+        .Package(url: "https://github.com/FabrizioBrancati/BFKit-Swift.git", majorVersion: 2, minor: 2)
     ],
     exclude: [
         "Config",