+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.<br>
 
 ---
 
+### 2.X Releases
+- `2.0.x` Releases - [2.0.0](#100)
+
 ### 1.X Releases
 - `1.2.x` Releases - [1.2.0](#120)
 - `1.1.x` Releases - [1.1.0](#110)
@@ -15,10 +18,21 @@ All notable changes to this project will be documented in this file.<br>
 
 ---
 
+## [2.0.0](https://github.com/FabrizioBrancati/SwiftyBot/releases/tag/2.0.0) - Sell Me Something
+### 07 Feb 2017
+### Added
+- Added Facebook Messenger support
+
+### Improved
+- Updated Vapor to 1.5.x
+- Updated BFKit-Swift to 2.2.x
+- Updated SwiftLint to 0.16.1
+
+
 ## [1.2.0](https://github.com/FabrizioBrancati/SwiftyBot/releases/tag/1.2.0) - General Update
 ### 23 Dec 2016
 ### Added
-- Added support for Xcode 8.2 and Swift 3.0.2
+- Added Xcode 8.2 and Swift 3.0.2 support
 
 ### Improved
 - Updated Vapor to 1.3.x
@@ -26,14 +40,14 @@ All notable changes to this project will be documented in this file.<br>
 - Updated SwiftLint to 0.14.0
 
 
-## [1.1.0](https://github.com/FabrizioBrancati/SwiftyBot/releases/tag/1.1.0) - BFKit Swift is awesome
+## [1.1.0](https://github.com/FabrizioBrancati/SwiftyBot/releases/tag/1.1.0) - BFKit Swift Is Awesome
 ### 10 Dec 2016
 ### Added
 - Added [BFKit Swift](https://github.com/FabrizioBrancati/BFKit-Swift) dependency
-- Added support for Xcode 8.1 and Swift 3.0.1
+- Added Xcode 8.1 and Swift 3.0.1 support
 
 
-## [1.0.0](https://github.com/FabrizioBrancati/SwiftyBot/releases/tag/1.0.0) - Send me a Telegram
+## [1.0.0](https://github.com/FabrizioBrancati/SwiftyBot/releases/tag/1.0.0) - Send Me A Telegram
 ### 17 Oct 2016
 ### Added
 - Added Telegram support